requirements.txt
pygame-ce
numpy
opencv-python
ffpyplayer
